requirements.txt
# Requisites
Click==7.0
PyYAML==5.3
requests==2.22.0
# Pre-requisites
certifi==2019.11.28
chardet==3.0.4
idna==2.8
urllib3==1.25.8
